'Verifier::visitInstruction' was done.
But we did not use 'ssreflect'. To connect the bool world and 
the logical world, for a particular relation, we must define both 
a boolean-version of this relation, and the Prop-version of this 
relation, and then use 'reflect' to connect them. Given a proposition, 
if the proposition is decidable, is there a way to generate its 
computational counterpart automatically? Coq doesn't allow us to 
destruct 'decidable' to return boolean, because boolean is in Set, 
while 'decidable'  is in Prop. I guess, this is how 'reflect' solved 
this problem? So I defined the two parts manually for each relation.

Ssreflect proved a lot of 'reflect' between logical connectives and 
the corresponding computational operations. say orP, with apply/orP 
and move/orP, we can switch between logical and computational. These 
apply/orP and move/orP tactics seem to be implemented by the extension 
of Coq core, not to be done with the existing Coq tactics.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2343 eef5ea85-cc2e-0410-bd84-0fae1f012da8