1) supported recursive types   a) to rule out 'infinite recursive types', such as       %1 = {%2}       %2 = {%1}      we need to define well-formedness of named type tables, s.t. there is no circular dependency. Therefore, the type size calculation can be defined as a simple recursion by passing the sorted type table in term of acyclic dependency.
      %1 = {%2*}
      %2 = {%1*}
     are good types, because the pointers stop calculating type size
infinitely, a pointer type always has a fixed size.

  b) Equations between recursive types need to be extended, because LLVM
uses structural types -- two types are equal if they have same
structure. 

  c) regression shows the errors of recursive types are removed. But
some cases run slower (indicated in Interpreter/report/report.lynx.txt)
for some reason...

2) testcases/softbound/*.ll are code for softbound.
   The extracted interpreter can execute softbound except that there
is one foreign wrapper function that uses inline asm. It is not used, so
we commented it out from vol/tests

   To support inline asm, we can provide interfaces that, given input
states and asm code, return output states in term of users' definitions.

   At this point, the formalism is ready to prove softbound.

3) fixed typos in README

4) checked in missing files of modified LLVM code. To avoid this
problem, we may need to check in the orginal 2.6 code into SVN, and
work with this SVN code. Because the size of the patch is getting too
large to maintain...

5) Interpreter/report/report.lynx.*** are regression results. 
   llvm-2.6/projects/test-suite/SingleSource/Regression/Lynx/ includes
the simple regression cases for the extracted interpreter, which were
copied from src/Interpreter/testcases
   


git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2720 eef5ea85-cc2e-0410-bd84-0fae1f012da8