1) redesigned mtv by recursion over a new syntax. This syntax extractly matches the structure of mtv, so we can do induction over the syntax.
This needs additional translations between the LLVM syntax and the new
syntax.

2) defined a new opsem over this new syntax, which also considers how to
handle signals like abort.

To fix
1) rewrite ssa_lib.v with a more generic data structure, then we can
reuse libraries for both the LLVM syntax and its invariant.
2) taking the last six instructions as an abstract does not always work
  milli/fptrret and milli/funcret2 are examples.
3) sub_tv_correct.v must be proved in a reversed way of eq_tv_correct.v,
namely back-simulation --- an output program is not stuck, then its
input isn't suck either. The other direction is not true because SB
instrumentation is conservative.
4) need to extend memory model for inttoptr. An LLVM pointer from int is
still defined if the int is indeed an address. So taking a pointer as a
block and offset pair does not work.
5) like the PLDI proofs, the proofs in mtv_correct.v should be based on
preservation and progress.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2844 eef5ea85-cc2e-0410-bd84-0fae1f012da8