1) Fixed rmap to be parial 2) simplified simulation relations, s.t. they consider both value_id and value_const cases, which reduces 1/3 size of the proofs
Checkpoint before another change.

Here is the problem. The SoftBound instrumentation of malloc is like
this:
%p = malloc i8, i32 (ptrtoint (i32* (inttoptr (i32 42 to i32*)) %to i32))
%b = bitcast i8* %p to i8*
%e = gep i8* %p, i32 (ptrtoint (i32* (inttoptr (i32 42 to i32*)) %to i32))

[NOTE: The ptrtoint and inttoptr things are optimized away by llvm-as
-disable-opt. From the log of llvm-as, I cannot find which pass it runs
to do such constant-fold. Maybe by default such folding should happen.
So we cannot create a case to show the problem.]

In term of the current memory model, if 42 is not a valid address for
any allocated blocks, inttoptr returns null, and the final reresult
before malloc is 0. Suppose the malloc turns 42 to be an allocated
address, the constant returns 42 after the malloc. This is inconsistent
with LLVM.

To fix it, we are intruducing a new value, Vint2ptr : int -> val. Given
an inttoptr i, if the i is within a good memory block b with ofs, it
returns Vptr b ofs, otherwise it returns Vint2ptr i. In such an way, we
turned int2ptr to be total w/o losing any information.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2886 eef5ea85-cc2e-0410-bd84-0fae1f012da8