Another problem: How to make ptr2int to be total.
The current ptr2int has two cases
 ptr2int (Vint2ptr i) = i // the easy case
 ptr2int (Vptr b ofs) = ..?

The later case searchs the mapping from block-ofs to addresses, which
can be None if b does not exist. For example
  p = malloc ...;
  free p;
  i = ptrtoint p; // Here, the p is not valid, so we are stuck.

For the current proof (proving a value is not changed before and after a
malloc), this does not seem to be matter. Because,
1) if ptrtoint is legal before malloc, it should returns the same thing
after malloc, because malloc does not break existing blocks
2) if ptrtoint is None by pointing to a dead block, it should still
point to the dead block since malloc does not reuse any old blocks.

The rest reasoning is to show that a well-formed program cannot forge a
ptr to a blk-ofs out-of-thin-air, which could be a hard proof.

But, even the above idea works, it is not extensible to a practical
memory model that reuses memory.
 
One idea is to split Vint to be a normal Vint and a Vptrtoint b ofs;
keeping track free-ed blocks. But we still need to argue that a program
cannot 'create' a ptr w/o malloc --- pointing to a block that is not
form malloc. Secondly, this is not extensible to reallocating malloc
either, because at the case, a blk can be mapped to more than one
concrete addresses. 

The other idea is to let Vptr to be
  blk -> ofs -> addr -> val
here, the addr is its concrete addr.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2897 eef5ea85-cc2e-0410-bd84-0fae1f012da8