checkpoint: the correctness of SBpass.
1) changed SBpass to make proofs
2) changed returnUpdateLocals, which should return result from callee to
caller even if caller does not need the result. The old design ignored
such return results, which is not true because that means the callee
knows what signature its callsite uses.

We did not find the problem before because the proofs did not need the
fact --- callee and caller may use different signatures, while the
SoftBound proofs does.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3009 eef5ea85-cc2e-0410-bd84-0fae1f012da8