1) checked in the regression logs
2) the r3416 also updated llvm2coq.ml: for the cmds that have no names
in term of syntax, say store, br, ret.., the version < 3415 assignes
them -1, following that LLVM does. Now, we assign them a unique negative
numbers as names. By doing this way, we can know which store to remove.

This also meets that Vellvm typechecker requires, all cmds (not only the
one that has id syntatically) must have uniq ids, and in proofs, we use them 
as names.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3417 eef5ea85-cc2e-0410-bd84-0fae1f012da8