id_rhs_val is proving that if we can show that id1 := rhs equals to v2 at a pc where id1 is defined (suppose v2 >> id1), then at any pc where both id1 and v2 dominate, we have [id1] = [v2].
las is proving LAS is correct, which is based on alive_store and
id_rhs_val.

Note: this proof intends to be more general than Lemma 17 that is
if id1 = rhs, and [rhs] = [v], then we can substitute id1 with v where both id1 
and v dominate. Lemma 17 is a relational property for a pair of
programs. Here, id_rhs_val extracts the invariant for a single program,
and could potentially be used in more contexts.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3622 eef5ea85-cc2e-0410-bd84-0fae1f012da8