confirmed that INV: "forall (v,chunk) in gv, v:chunk" is the invariant to prove store;load is identical.
In CompCert, there are similar lemmas. The lemmas are used by spilling
pass, and stack records layout pass where they need to show register
spilling and reloading is identical. In the CompCert Memory Model, we
cannot prove that
  if store ptr v; r = load ptr, then [r] = [v]
For example, when v is 0xffff of type int8, r is 0xff, so, we only have 
  [r] = (t)[v] where ptr:t*
This is not suprising, because memory st/ld pair is an implicit
conversion. 

In CompCert, the spilling pass does not need to prove [r]=[v] because
the target language uses an 'abstract' stack that proves this property
for free. They need to prove this in the stack layout pass. At that
pass, CompCert concretly uses i32 chunks for any supported int (i8, i16,
i32) values n CompCert; and double slots for both float and double
values. Luckily, (t)[v] = [v] when t = i32 or double in CompCert, so,
the property holds. I think this is because C spec does not specify how
many bytes to hold i8 and i16 values on stack. (Note, when t=float, 
double-single precision conversion is needed).

LLVM is byte-oriented, and requires the number of bytes to hold a value
is the minimal bytes to contain the corresponding bitsize up to
rounding. So, we need a more precise low-level t than CompCert---chunk
that relates values' size and types' bitsize, and show that the opsem preserves 
INV.  Then we should have:
  if ptr:chunk_of(t*), v:chunk_of(t), and store ptr v; r = load ptr, 
  then [r] = [v].

NOTE: this does not mean LLVM is sound. Because undef can be of any
chunk.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3817 eef5ea85-cc2e-0410-bd84-0fae1f012da8