1) solved the 2nd problem in r3863 in an ad-hoc way.
mem2reg only optimizes store-load pairs locally within blocks. So, we
only need to extend the domination relation to support instructions that
do not define variables (say, store, call void, ...) locally. This changed
metatheory in the minimal way, but is sufficient to finish the
current proofs.

In the future work when we are verifying global (across blocks) load/store
transformations, we definitely need to extend domination relations
completely (i.e. the inscope_of_id/cmd/tmn relations). But, I will leave this to
 the future me.

2) The next is to solve the last issue about domination-relative lemmas
in semantics-preservation proofs. Typing rules only ensure domination
relations for reachable blocks (which contain the uses of variables),
because dominator analysis is only sound to reachable blocks. 
So, for example, the corrrectness of LAS pass in an unreachable block
should not reply on domination properties. We may want to show
transformations in unreachable blocks are always correct (as long as
they preserve well-formnedss) since they do not execute.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3867 eef5ea85-cc2e-0410-bd84-0fae1f012da8