1) Proved all diverging simulations.
NOTE: Coinductive lemmas are not always type-checkable. Suppose we have
 a coinductive definition:
   a --> oo iff a --> a' /\ a' --> oo
 and we want to prove (Lem 1)
   P a b /\ b --> oo ==> a --> oo
 b --> oo gives b --> b' /\ b' --> oo
 supposed, we have
   P a b /\ b --> b', then ex a', a -->+ a' /\ P a' b'
 By coinduction, we have
   a' --> oo

 But, even if we can prove (Lem 2)
   a -->+ a' /\ a --> oo ==> a --> oo
   (this can be proved both inductively or coinductively),
 This (Lem2) cannot conclude the lemma 1, since the use of coinductive 
 hypothesis is not under a coinductive constructor directly.

 One way to solve the problem is to 'inline' the proofs of Lem2 in Lem1,
and let the coinductive hyp used 'close' to the co-constructor, s.t. Coq
can type-check it. We applied the idea in some proofs. But, this could make 
proofs large and complex.

 The other way is by defining a helper inductive definition
    a ==> oo iff a -->+ a' /\ a ==> oo
 Lem3 P a b /\ b ==> oo ==> a ==> oo can be proved, because the -->+ in
==>oo rules let us use coinductive hyp under constructor directly.

 Then we can show ==>oo iff -->oo co-inductively, which proves
Lem1. This idea can lead to simpler proofs, and is used in most of our proofs.
  
2) Extraction is broken from r3899.

In r3899, the LLVMsyntax module before r3899 is moved into
LLVMsyntax_base in r3899, the LLVMsyntax in r3899 includes
LLVMsyntax_base.

But, it seems that Coq extraction does not work well for module
inclusion. For example, Extraction/extraction_core.v < r3899 maps data
types in LLVMsyntax into OCaml types. But, now, if we extract
both LLVMsyntax_base/LLVMsyntax into OCaml types, the syntax.mli still takes 
the old types defined in Coq. For some reason, Coq extraction will ignore the 
mapping for LLVMsyntax. Because module inclusion means all types in LLVMsynax 
= all types in LLVMsyntax_base, and this equation might overwrite our 
extraction, So Coq does nothing for LLVMsyntax.

If the problem cannot be resolved, we may have to roll back to the
infrastructure < r3899.

3) The Makefile.coq generated from Make added one more / for all files
Vellvm, such as
 Vellvm//syntax.v ..
this breaks all the dependences. Files in other directories are
fine. I did not find any place where the / was added.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3900 eef5ea85-cc2e-0410-bd84-0fae1f012da8