resolved the extraction issue 1) the problem could be that Coq extraction does not work when including a module with nested module into another module, For example
  Module A.
    Module a.
      Definition t:=nat.
      ...
    End a.   
    Definition s:=...
  End A.

  Module B.
    include A.
  End B.

After extraction, the OCaml code is like

  module A =
    module a =
      var t = ...
    end
 
    var s = ...
  end

  module B =
    module a =
      var t = ...
    end
 
    var s = A.s
  end

The nested module a is copied, while the s is constrainted by an
equation. If we extract A.a.t to be Int, then in OCaml, A.a is Int, but
the B.a is still nat. Extracting B.a.t is ignored for some reason.
Extraction A.s works good.

To walk around the issue, we define the nested module into
Vellvm/datatype_base.v, which, therefore, defines every module on the
top-level. Maybe this is not an issue for Coq extraction, if any nested
module can be rewritten in such a way??

2) modified OCaml binding files to work with new data types w/o mutual
inductive definitions.

3) The proof of list_value_dec (>=r3899) in Vellvm/infrastructure.v uses
eq_nat_dec for proving decidability of Size.t. Although Size.t = nat, on
extraction, Size.t and Size.dec are extracted to int, and int's cmp,
while nat and eq_nat_dec are unchanged. So, we need to use Size.dec in
the proof.

4) Extracted interpreters and mem2reg will work. We had not changed the
their computing part (>=r3899).



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3963 eef5ea85-cc2e-0410-bd84-0fae1f012da8