for a cmd in block b that uses id0, if the b is not reachable, typings do not need to check any property of id0, because the b cannot execute at runtime. The old rule still checked if id0 is defined in some block in the unreachable case. This complicated substitution-preserving-WF proofs.
In LAS, we require the stored value dominates the loaded temporary only
when the load's block b is reachable, but allow anything when b is
unreachable. By doing so, the algorithm is simpler w/o checking
reachability. But, in the old typing rules, when b is unreachable, we
still need to find the block that defines the stored value. Although
this is provable, the proof is not clean. So, we changed the wf_operand
rule to check nothing when a block under check is unreachable.

Save point before the same update for phinode.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3966 eef5ea85-cc2e-0410-bd84-0fae1f012da8