1) The proof of filter (r3969) was correct, but not strong enough at all cases. Filter proved that if a function does not use the variable defined in a pure cmd of the function, then we can remove the cmd. Here is a case,
mem2reg calls DIE after LAS to remove the load, because all uses of the variable
 lid defined by the load were already substed. But, to conclude that no one
uses lid, we need to show that the substituting value v (in this case,
which is the value stored alive to the load) does not contain lid. If
the store is in a reachable block, this is true, otherwise, v can
contain lid. So, in term of the 3969's DIE, the lid is not removable.

Actually, it is fine to remove a variable if the variable is not used
only when the variable is defined in a reachable block. If the variable
to remove is in an unreachable block, other unreachable blocks can still
use it. In the case, it is still safe to remove, because unreachable
blocks have no effects. So, we strengthened die.v (simulation) in this way. 
LAS's output program holds the new condition.

The alternative is that mem2reg should check the condition explicitly.
But, this introduces compile-time cost. 

die_wfS.v is broken. We need to strenthen filter.v to meet the same
condition.

2) Note that this statement in r3969 was already an extension of 
Vminus/removal.v that does not require that there exists a cmd in the
function that defines the variable. Vminus/removal.v is not sufficient to prove
 that removing a cmd whose id equals to the id of a dead store is safe.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3970 eef5ea85-cc2e-0410-bd84-0fae1f012da8