1) Resolved the 1st issue in r3971: how to check unreachable blocks. Actually, the current Vellvm type checker is consistent with LLVM's.   1) ignore checking block/instruction-level domination in unreachable blocks   2) check if a used variable defined in unreachable blocks. Ex1:
define i32 @foo() nounwind {
entry:
  %p = alloca i32, align 4
  ..
  ret i32 42

dead:
  store i32 %x, i32* %p
  %x = load i32* %p           ; good
  %y = add i32 %t, %x          ; bad
  unreachable
}

In dead, LLVM does not check the store uses a non-dom %x, but check that
%t is undefined yet.
Ex2:

define i32 @bar() nounwind {
entry:
  %p = alloca i32, align 4
  ..
  ret i32 42

dead1
  store i32 %x, i32* %p       ; good
  unreachable

dead2:
  %x = load i32* %p           ; good
  unreachable
}

In this case, it is also good.

The reason LLVM ignores domination checking in dead blocks is that
dom-analysis is unsound w.r.t. unreachable blocks. 

But, ideally, we can only ignoring block-level checking (across blocks,
See Ex2).
1) In Ex1, instruction-level domination (within blocks) is sound no matter 
whether dead or not. Reporting nothing is fine, since the code is dead.
2) In Ex1, we do not need to check if %t is defined. This warning
disallows some transformations, and forces passes to clear all
dependent dead code recursively. 

We decided to make the Vellvm typing rules unchanged, which is
consistent to what LLVM does. For the problem (the first issue in
r3970), we have at least two ways to go.
1) after deleting lid, we should recurively remove all dependent code
from dead blocks. This will require more proofs.
2) Only do LAS in reachable code. In reachable blocks, the
sefl-reference in Ex1 cannot happen. Also, the phinode placement pass
only inserted additional loads/stores in reachable blocks. So, doing so
will generate the same result.

We took the same way. Experimentally, the code generated is equivalent
to the old one. In other words, we do not remove less loads/stores by
only LAS reachable blocks.

NOTE: 
1) the original LLVM's mem2reg only inserts phinodes and rename variables in 
   reachable blocks too. 
2) Another observation, recall the pruned SSA problem (See r3717). I
think the standard DIE should remove the rest phinodes, because at most
of cases, they are dead forwardly. Pruned SSA does not allow to insert
phinode in the following case:

  if c then 
    y = 1 
  else 
    y = 2
  x = y + 1
  ret 0

because agressively the use of y is dead (ret doest not use). But, this
may not always be the case in practice. Experimentally the performance
gap between wmem2reg and mem2reg is vert small even w/o the deletion.

If we really need this, we should extend the current die.v to alow
removing phi-nodes, which must be a by-product of the phi-node elimination
pass.

2) At this commit, the entire code under src3.0 contains 4 unproved
lemmas. 3 of them are from the 2nd/3rd issues in r3971. The rest one is
about the phinode-elimination pass.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3974 eef5ea85-cc2e-0410-bd84-0fae1f012da8