1) designed new OCaml bindings for validating LLVM's dominator tree algorithm.
NOTE: When adding a pass (say domtree) into a LLVM Function Pass Manager, the
manager actually works on a copy of the pass, although it still
correctly deletes the original pass then it was disposed. I had planned
to use Function Pass to expose the dom-tree interface. But actually
exposing them directly is simpler because of the above technical issue.

2) The Vellvm/Dominators is the working directory for CPP12. Eventually
this directory will define the interfaces and instances of dom-analysis
that Vellvm should be refractored to use.

3) There is a technical question: where the validated dom-tree should be
stored. One idea could be storing the tree into AST with its function.
But this will change proofs. An alternative could be storing the trees
"globally", and assuming there is way that we can fetch by function
names. The choice wont affect correctness proofs of validation and
experiments, which is only a problem when we want to use the validated
dtree in proofs.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3986 eef5ea85-cc2e-0410-bd84-0fae1f012da8