** Implemented a fast, but simple (both to implement and to prove) dominitor analysis, following the idea from   A Simple, Fast Dominance Algorithm   by Keith D. Cooper , Timothy J. Harvey , Ken Kennedy
The performance result w.r.t spec95 and olden matches the figure 2 in
  Finding dominators in practice
  by: Loukas Georgiadis, Renato F. Werneck, Robert E. Tarjan, David I. August

1) for all the benchmarks, the fast dom-analysis has the same
performance of LLVM's, which uses a complex and fast (E (log N)) algorithm
  A fast algorithm for finding dominators in a flowgraph
  by: Thomas Lengauer, Robert E. Tarjan

2) The native dom-analysis (based on Aho and Ullman) is much slower than them.
  bh: 6x
  tsp: 2x
  129.compress: 2x
  099.go: 22x
  130.li: 6x
  132.jpeg: 7x
  
  In spec2000 and spec2006, there are more benchmarks similar to 099.go.

** The fast dom we implemented in Coq is different than the one
published in the following aspects

1) We implemented the algorithm by extending the Kildall module in Coq.
  a) The original algorithm picks a node, and then propagates
information from all its predecessors. Kildall picks a node, then
propagrates its information to all its successors. 

  b) The existent Kildall requires that the content of each node is a
lattice, which must be proved w/o looking into the algorithm. But in the
new dom-analysis, this can only be proved when we consider the invariant
of all nodes' contents. So, we weakened Kildall module to allow this.

2) The original algorithm uses a map to stores the idom (known so far)
for each node. We maintain a map from node to a list of all strict
dominators (the list is ordered by idom relation). Note that we do not
need to sort and maintain the order explicitly, like the original
algorithm, we should prove that the order in which the algorithm
discovers dominators is exactly the order, which is a property the
original algo should hold too.

  We hope the design can simplify the proofs. Consider a CFG with
N nodes and E edges. For each iteration, the algorithm needs to visit
all (in the worst case) edges once. Suppose the average size of
dominators is log(N). Then, when visiting each node, the original algo
needs log(N) * log(N) data accesses (each one is a tree lookup), we need 
  log(N) + log(N) - 1 data accesses (one tree lookup, others are list
traversals). 

After the computation, the data free/allocation by OCaml is N in the
original, N * log(N) in ours. Let us say A is the cost of data access, B
is the cost of data allication. Then, in each iteration it is 
  A((logN)^2) + BN vs A(2logN-1) + B(N*log(N))

However, the intersect operation in our algorithm can be written down
directly, since it is like a traversal of a finit list. But, the
original algorithm's is terminating only if we can prove all the
reachable nodes are finite.

3) This was a technique issue to make the algorithm work. We actually
need an efficient node representation that can make map/tree search
efficient. So far, nodes are named by String.t. When creating postorder
of nodes, we created binary (positive in Coq) to represent them, so we
can even use radix tree to build a search map/tree.

We may want to apply the idea to Vellvm: when converting C++ AST to Coq
AST, for each function, we should generate the postorder positive
numbers to each node by DFS, which should improve performance a lot.
String-based names will only be used to pretty-printing.

4) Why the old dom-analysis slow?

  a) it uses String.t to represent nodes, which introduces comparison
     costs, and does not allow radix tree.
  b) it does not iterate CFG in the reserved post-order
  c) it does not check redundant elements in sets, which takes
     allocation/searching/intersection costs
  d) it has to check if two sets equal, and intersect two sets.

5) No correctness proofs of the new algorithm were done. DFS is
implemented in OCaml, it is not obvious terminating in Coq.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3989 eef5ea85-cc2e-0410-bd84-0fae1f012da8