Proved that fast-dom algorithm is monotonic. Found a bug of the implementation in the proofs:   transfer n Dom[n] should equal to None when Dom[n] = None, since we   take None as the whole set. The old design assigned n::nil in the case,   which could generate wrong results: say we pick uninitializied node n   whose value is None, then all its successors will be assigned to n::nil,   eventually will be nil when they are intersected with a path from entry.
However, the reversed post order we pick prevented the error for
reachable nodes. DFS gives us that a reachable node must have at least
one predecessor whose rev-postorder is larger than the node's.
Intuitively each path on a DFS tree is a simpl path from root of the original
CFG, and at least reachable nodes' predecessor must be on the the simpl
path. Given the fact, we know if we always pick larger rev-postorder
first, we will never pick a node all of whose preds have not been picked; and
we know once a node is visited, all its succs are initialized;
plus, entry must be picked first. So, we will never pick an
uninitialized reachable node in the buggy algorithm. This explained why
we did not observe the bug.

But, for unreachable code, the algorithm for sure will generate wrong
results. By definition, unreachable nodes should be stricly dominated by
all others. The bug will eventially only assign the node self as its
strict dominator incorrectly. We did not check unreachable nodes, so
missed the bug.

We added a log/dom-debug.log as a regression test for dom-trees.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3994 eef5ea85-cc2e-0410-bd84-0fae1f012da8