****** 1) modified the implementation to be rev-postorder (RPO) + pulling information    from predecessors. 2) designed a new variant: breadth-first search order (BFS) + pushing    information to successors. Because...
r3989 discussed one difference between our implementation and the
original 
algorithm.
  The original algorithm picks a node, and then (pull) propagates
information 
  from all its predecessors. We pick a node, then (push) propagrate its 
  information to all its successors.

It turned out that this design is not easier to prove correct, although
it does 
not seem to be wrong either. Consider a CFG,
 
   +---+----------+
   |   |          |
   |   \/         \/
   4 => 3 => 2 => 1

Here => denotes a DFG tree edge (which is also a CFG edge), -> denodes a
non-tree (but CFG) edge. For the correctness of the algorithm, we want
to prove 
that
  INV: all uninitialized nodes should be of a smaller number that
initialized 
  ones. 
On the CFG, after 4 is processed, 3 and 1 (which are successes of 4) are 
initialized. 2 is not initialized yet, but is larger than 1.

Actually, no matter what order the algorithm chooses, the algorithm
should be 
correct w.r.t a total order TO that is created by the order a node
initialized.
In this case, the TO is 4 3 1 2. We should still be able to prove INV by 
replacing RPO by TO. 

The original algorithm (although not explained explicitly) uses pull +
RPO to
force TO = RPO. Our modified algorithm does push + RPO, which <> RPO. In
fact,
push + BFS = TO. Here is the BFS of the above CFG.

   +---+----------+
   |   |          |
   |   \/         \/
   4 => 3 => 1 --> 2
   ||             /\
   ||             || 
   +===============+  

So, we implemented two both of them. BFS is implemented in OCaml in this
version.

Experiments showed that push+BFS is always faster than pull+RPO (1X-4X), 
although for larger programs, both of them are much faster than the
native-one, 
and close to LLVM's.

pull+RPO is slower than push+BFS because
1) pull processes more back-edges (from child to ancestor) on DFS, which
does
   not contribute to dom-analysis at all
2) for a n and its pred p, to recompute IN(n), we need to either compute
OUT(p)
   from IN(p) or pre-compute OUTs and carry them in arguments.
   In the first case, we need to compute E (# of edges)'s OUT per each
iteration,
   in the second case, we need another N copies for each iteration
  
   In push mode, given a n, we compute its OUT, and then propogate to
all its
   sucessors, each iteration only needs N's OUT computation.
3) In push mode, we can directly know which successors are changed, and
only 
   add them to worklist. But in pull mode, when n is changed, we have to
add
   all of its successors to worklist.
4) Finding dominators in practice
    by: Loukas Georgiadis, Renato F. Werneck, Robert E. Tarjan, David I.
August
   also observed that BFS+pull is better than RPO+pull in imperative
setting.
   I think in Coq, BFS+push should be better than BFS+pull.

   Section 3 of the paper very briefly discussed that in all 5 kinds of
worst 
   cases, BFS wins RPO in 4 out of 5.

Since the proofs of them are similar. We may prefer to push+BFS.

*****
Added spec2000 and spec2006 into benchmarks. See log/time.log. All the
results
are consistent with spec95's results. In particular, BFS/native
458.sjeng: 188x
177.mesa: 228x
188.ammp: 41x
...

186.crafty fails because it includes inclined asm that Vellvm does not
support.

log/bfs.log and log/rpo.log are regression results.

******
In Dominators, split fast_iter.v tobe
  dom_libs.v: basic definitions
  dfs.v: terminating DFS
  rpo_iter.v: pull+rpo
  bfs_iter.v: push+bfs



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3995 eef5ea85-cc2e-0410-bd84-0fae1f012da8